refactor: move hub into features
