+++ client/src/features/hub/pages/EmailDetail.tsx
